Transactions

+++ database/migrations/2025_03_03_131313_create_transaction_details_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('transaction_id')->constrained('transactions');
             $table->foreignId('product_id')->constrained('products');
             $table->integer('quantity');
-            $table->decimal('subtotal', 10, 2);
+            $table->integer('subtotal');
             $table->timestamps();
             $table->softDeletes();
         });
